# frontend/public/data/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Anuvaad API",
    "description": "**Anuvaad** \u2014 AI-powered code translation platform.\n\nTranslate code to plain English, English to code, and between 35+ languages.\n\nAll new clients should use the `/api/v1/` prefix. Legacy `/api/` routes are deprecated and will be removed on **2027-01-01**.",
    "contact": {
      "name": "Anuvaad Support",
      "url": "http://localhost:3000/"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    },
    "version": "3.0.0"
  },
  "paths": {
...
    "/api/v1/oauth/github/callback": {
      "post": {
        "summary": "Github Callback",
        "description": "Handle GitHub OAuth callback \u2014 exchange code for token.\n\nFIX-01: Token is Fernet-encrypted before DB storage.\nFIX-26: Uses ORM repository instead of supabase_request().\nSEC-GH-01: Plaintext access_token is no longer returned in the response body.\n           The token is stored encrypted in the database; subsequent API calls\n           (e.g. GET /github/repos) transparently use the stored token.\nSEC-GH-02: Rate-limited to 5 requests/min per token to prevent abuse.",
        "operationId": "github_callback_api_v1_oauth_github_callback_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "code",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Code"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/github/repos": {
      "get": {
        "summary": "Get Github Repos",
        "description": "Return the GitHub repositories for the authenticated user.\n\nFIX-26: ORM-based token retrieval. FIX-30: Removed redundant auth guard.",
        "operationId": "get_github_repos_api_v1_github_repos_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/github/process-repo": {
      "post": {
        "summary": "Process Github Repo",
        "description": "Enqueue a background task to process a GitHub repository.\n\nFIX-30: Removed redundant auth guard.\nSEC-GH-03: Validates repo_name format before enqueueing.",
        "operationId": "process_github_repo_api_v1_github_process_repo_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "repo_name",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Repo Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/github/disconnect": {
      "delete": {
        "summary": "Disconnect Github",
        "description": "Remove the stored GitHub OAuth token for the authenticated user.",
        "operationId": "disconnect_github_api_v1_github_disconnect_delete",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
    "/api/oauth/github/callback": {
      "post": {
        "summary": "Github Callback",
        "description": "Handle GitHub OAuth callback \u2014 exchange code for token.\n\nFIX-01: Token is Fernet-encrypted before DB storage.\nFIX-26: Uses ORM repository instead of supabase_request().\nSEC-GH-01: Plaintext access_token is no longer returned in the response body.\n           The token is stored encrypted in the database; subsequent API calls\n           (e.g. GET /github/repos) transparently use the stored token.\nSEC-GH-02: Rate-limited to 5 requests/min per token to prevent abuse.",
        "operationId": "github_callback_api_oauth_github_callback_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "code",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Code"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/github/repos": {
      "get": {
        "summary": "Get Github Repos",
        "description": "Return the GitHub repositories for the authenticated user.\n\nFIX-26: ORM-based token retrieval. FIX-30: Removed redundant auth guard.",
        "operationId": "get_github_repos_api_github_repos_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/github/process-repo": {
      "post": {
        "summary": "Process Github Repo",
        "description": "Enqueue a background task to process a GitHub repository.\n\nFIX-30: Removed redundant auth guard.\nSEC-GH-03: Validates repo_name format before enqueueing.",
        "operationId": "process_github_repo_api_github_process_repo_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "repo_name",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Repo Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/github/disconnect": {
      "delete": {
        "summary": "Disconnect Github",
        "description": "Remove the stored GitHub OAuth token for the authenticated user.",
        "operationId": "disconnect_github_api_github_disconnect_delete",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}